feat: allow floats for CPU cloud-spec values (#199)
* support floats for cpus and redis storage to allow for a more fine
grained mapping of resources to "CPU"

* assign missing client factory

* process constructor

* rename Epsilon to Delta

* move files

* reformat json

* update schema to support floating point numbers

* revert allowing float64 for redis as it is not needed

* remove unnecessary field

* add logger

* Revert "add logger"

This reverts commit d318149da027297cfc4668fc035753795f0decd0.

+++ pkg/config/public_cloud_specs.go
@@ -30,7 +30,7 @@ type Providers struct {
 }
 
 type Feature struct {
-	CpuCores int     `json:"cpu_cores"`
+	CpuCores float64 `json:"cpu_cores"`
 	Memory   float64 `json:"memory"`
 }
